Make viewonetopic component reusable - Create a compnent - Move the comment component into the view one component - Refator the viewone to accommodate the changesy
