+++ src/components/TopicOneView/index.js
@@ -1,0 +1,1 @@
+export * from './TopicOneView';
