+++ src/components/index.js
@@ -3,11 +3,8 @@ export * from './Radio';
 export * from './TopicsCarousel';
 export * from './FileUploader';
 export * from './NotFound';
-export * from './CommentaryForm';
 export * from './AdminCommuniques';
-export * from './Comments';
 export * from './Commentaries';
-export * from './CommentaryForm';
 export * from './AudioPlayer';
 export * from './Page';
 export * from './MenuSetting';
